Support STR pre-amplifier.
Only one zone.
Only four listening modes.
Attenuation range from -96dB to +7dB in 0.5dB steps.
The BRT command, which means query bitrate in most devices, is changed
to move the left/right balance to the right.  Z1BRT0 is equivalent to
moving balance 0.5dB to the right, Z1BRT50 moves it by 2.5dB and
Z1BRT100 by 5dB.  When we send Z1BRT? to query bitrate, we actually shift
the balance 0.5dB to the right, which is probably a bug on the device.
Simply using this API would inadvertently adjust the balance fully to
the right over time!